Modified file

+++ src/loadPCD.cpp
@@ -6,7 +6,6 @@
 #include <opencv2/core/core.hpp>
 #include <opencv2/highgui/highgui.hpp>
 #include <libfreenect/libfreenect.h>
-#include <pcl/opencv_grabber.h>
 
 
 
@@ -15,17 +14,16 @@ using namespace cv;
 bool loadData(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud);
 
 int main(){
+	
 	//skapar ett "rum"
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud (new pcl::PointCloud<pcl::PointXYZ>);
-	//laddar filen
+	
+	//laddar filen, read from disc
 	if(!loadData(cloud)){
 		std::cout << "failed to load file" << std::endl;
   		return 0;
 	}
 	
-
-
-
 	return 0;
 }
 
@@ -37,10 +35,11 @@ bool loadData(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud){
 		//PCLERROR("no file");
 		return false;
 	}
+	
 	std::cout << "file size " << std::endl;
 	std::cout << "X:" << cloud->width << '\t' << "Y:" << cloud->height << std::endl;
 
-	for (size_t i = 0; i < cloud->points.size (); ++i)
+	for (int i = 0; i < cloud->points.size (); ++i)
 	{
 	 	  std::cout << "    " << cloud->points[i].x
               << " "    << cloud->points[i].y
